Rename Plugin N to P to avoid collision with another Plugin N
