+++ ReferencePluginP/packages.config
